Build a HIBAG model

To build a HIBAG model for predicting HLA types.

hlaAttrBagging(hla, snp, nclassifier=100, mtry=c("sqrt", "all", "one"),
  prune=TRUE, rm.na=TRUE, verbose=TRUE, verbose.detail=FALSE)

Arguments

hla

the training HLA types, an object of hlaAlleleClass

snp

the training SNP genotypes, an object of hlaSNPGenoClass

nclassifier

the total number of individual classifiers

mtry

a character or a numeric value, the number of variables randomly sampled as candidates for each selection. See details

prune

if TRUE, to perform a parsimonious forward variable selection, otherwise, exhaustive forward variable selection. See details

rm.na

if TRUE, remove the samples with missing HLA types

verbose

if TRUE, show information

verbose.detail

if TRUE, show more information

Details

mtry (the number of variables randomly sampled as candidates for each selection): "sqrt", using the square root of the total number of candidate SNPs; "all", using all candidate SNPs; "one", using one SNP; an integer, specifying the number of candidate SNPs; 0 < r < 1, the number of candidate SNPs is "r * the total number of SNPs".

prune: there is no significant difference on accuracy between parsimonious and exhaustive forward variable selections. If prune=TRUE, the searching algorithm performs a parsimonious forward variable selection: if a new SNP predictor reduces the current out-of-bag accuracy, then it is removed from the candidate SNP set for future searching. Parsimonious selection helps to improve the computational efficiency by reducing the searching times on non-informative SNP markers.

A parallel version of hlaAttrBagging is hlaParallelAttrBagging.

Value

Return an object of hlaAttrBagClass:

n.samp

the total number of training samples

n.snp

the total number of candidate SNP predictors

sample.id

the sample IDs

snp.id

the SNP IDs

snp.position

SNP position in basepair

snp.allele

a vector of characters with the format of “A allele/B allele”

snp.allele.freq

the allele frequencies

hla.locus

the name of HLA locus

hla.allele

the HLA alleles used in the model

hla.freq

the HLA allele frequencies

assembly

the human genome reference, such like "hg19"

model

internal use

References

Zheng X, Shen J, Cox C, Wakefield J, Ehm M, Nelson M, Weir BS; HIBAG – HLA Genotype Imputation with Attribute Bagging; (Abstract 294, Platform/Oral Talk); Present at the 62nd Annual Meeting of the American Society of Human Genetics, November 9, 2012 in San Francisco, California.

Zheng X, Shen J, Cox C, Wakefield J, Ehm M, Nelson M, Weir BS; HIBAG – HLA Genotype Imputation with Attribute Bagging. Pharmacogenomics Journal. doi: 10.1038/tpj.2013.18. http://dx.doi.org/10.1038/tpj.2013.18

Author

Xiuwen Zheng

See also

Examples

# load HLA types and SNP genotypes
data(HLA_Type_Table, package="HIBAG")
data(HapMap_CEU_Geno, package="HIBAG")

# make a "hlaAlleleClass" object
hla.id <- "A"
hla <- hlaAllele(HLA_Type_Table$sample.id,
  H1 = HLA_Type_Table[, paste(hla.id, ".1", sep="")],
  H2 = HLA_Type_Table[, paste(hla.id, ".2", sep="")],
  locus=hla.id, assembly="hg19")

# divide HLA types randomly
set.seed(100)
hlatab <- hlaSplitAllele(hla, train.prop=0.5)
names(hlatab)
#> [1] "training"   "validation"
# "training"   "validation"
summary(hlatab$training)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> # of unique KIR3DL1/S1 genotypes: 23
summary(hlatab$validation)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 26
#> # of unique KIR3DL1/S1 alleles: 12
#> # of unique KIR3DL1/S1 genotypes: 14

# SNP predictors within the flanking region on each side
region <- 500   # kb
snpid <- hlaFlankingSNP(HapMap_CEU_Geno$snp.id, HapMap_CEU_Geno$snp.position,
  hla.id, region*1000, assembly="hg19")
length(snpid)  # 275
#> [1] 275

# training and validation genotypes
train.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  snp.sel=match(snpid, HapMap_CEU_Geno$snp.id),
  samp.sel=match(hlatab$training$value$sample.id, HapMap_CEU_Geno$sample.id))
test.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  samp.sel=match(hlatab$validation$value$sample.id, HapMap_CEU_Geno$sample.id))

# train a HIBAG model
set.seed(100)
# please use "nclassifier=100" when you use HIBAG for real data
model <- hlaAttrBagging(hlatab$training, train.geno, nclassifier=4,
  verbose.detail=TRUE)
#> 11 monomorphic SNPs have been removed.
#> Build a PONG2 model with 4 individual classifiers:
#> # of SNPs randomly sampled as candidates for each selection: 17
#> # of SNPs: 264, # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> 	1  , added snp: 211, loss: 186.723, out-of-bag acc: 56.25%, # of haplo: 12
#> 	2  , added snp: 114, loss: 149.24, out-of-bag acc: 62.50%, # of haplo: 12
#> 	3  , added snp: 143, loss: 112.579, out-of-bag acc: 68.75%, # of haplo: 12
#> 	4  , added snp: 111, loss: 79.1564, out-of-bag acc: 71.88%, # of haplo: 12
#> 	5  , added snp: 227, loss: 53.4961, out-of-bag acc: 81.25%, # of haplo: 13
#> 	6  , added snp: 232, loss: 39.6983, out-of-bag acc: 84.38%, # of haplo: 13
#> 	7  , added snp: 150, loss: 33.27, out-of-bag acc: 84.38%, # of haplo: 13
#> 	8  , added snp: 263, loss: 32.7043, out-of-bag acc: 84.38%, # of haplo: 16
#> 	9  , added snp: 73, loss: 23.3549, out-of-bag acc: 84.38%, # of haplo: 18
#> 	10 , added snp: 151, loss: 23.0126, out-of-bag acc: 84.38%, # of haplo: 17
#> 	11 , added snp: 100, loss: 22.9579, out-of-bag acc: 84.38%, # of haplo: 18
#> 	12 , added snp: 226, loss: 22.9177, out-of-bag acc: 84.38%, # of haplo: 18
#> 	13 , added snp: 229, loss: 13.9143, out-of-bag acc: 84.38%, # of haplo: 18
#> 	14 , added snp: 199, loss: 10.8838, out-of-bag acc: 84.38%, # of haplo: 19
#> Tue May 19 19:47:18 2026,   1 individual classifier, out-of-bag acc: 84.38%, # of SNPs: 14, # of haplo: 19
#> 	1  , added snp: 191, loss: 189.361, out-of-bag acc: 75.00%, # of haplo: 13
#> 	2  , added snp: 155, loss: 152.12, out-of-bag acc: 81.25%, # of haplo: 13
#> 	3  , added snp: 109, loss: 109.822, out-of-bag acc: 81.25%, # of haplo: 13
#> 	4  , added snp: 123, loss: 64.6381, out-of-bag acc: 87.50%, # of haplo: 13
#> 	5  , added snp: 111, loss: 59.174, out-of-bag acc: 93.75%, # of haplo: 13
#> 	6  , added snp: 135, loss: 39.4287, out-of-bag acc: 93.75%, # of haplo: 13
#> 	7  , added snp: 205, loss: 26.0626, out-of-bag acc: 93.75%, # of haplo: 15
#> 	8  , added snp: 81, loss: 16.3915, out-of-bag acc: 93.75%, # of haplo: 18
#> 	9  , added snp: 48, loss: 12.3983, out-of-bag acc: 93.75%, # of haplo: 21
#> 	10 , added snp: 199, loss: 9.3508, out-of-bag acc: 93.75%, # of haplo: 21
#> 	11 , added snp: 216, loss: 5.81539, out-of-bag acc: 93.75%, # of haplo: 24
#> 	12 , added snp: 151, loss: 3.27629, out-of-bag acc: 93.75%, # of haplo: 25
#> 	13 , added snp: 236, loss: 1.89, out-of-bag acc: 93.75%, # of haplo: 29
#> 	14 , added snp: 73, loss: 1.41482, out-of-bag acc: 93.75%, # of haplo: 32
#> Tue May 19 19:47:18 2026,   2 individual classifier, out-of-bag acc: 93.75%, # of SNPs: 14, # of haplo: 32
#> 	1  , added snp: 181, loss: 189.371, out-of-bag acc: 60.71%, # of haplo: 13
#> 	2  , added snp: 118, loss: 171.486, out-of-bag acc: 67.86%, # of haplo: 13
#> 	3  , added snp: 146, loss: 158.952, out-of-bag acc: 75.00%, # of haplo: 13
#> 	4  , added snp: 137, loss: 122.21, out-of-bag acc: 78.57%, # of haplo: 13
#> 	5  , added snp: 132, loss: 76.1267, out-of-bag acc: 82.14%, # of haplo: 13
#> 	6  , added snp: 83, loss: 59.2983, out-of-bag acc: 85.71%, # of haplo: 15
#> 	7  , added snp: 158, loss: 37.9786, out-of-bag acc: 92.86%, # of haplo: 15
#> 	8  , added snp: 21, loss: 23.3557, out-of-bag acc: 92.86%, # of haplo: 16
#> 	9  , added snp: 47, loss: 22.6795, out-of-bag acc: 92.86%, # of haplo: 16
#> 	10 , added snp: 191, loss: 15.7477, out-of-bag acc: 92.86%, # of haplo: 16
#> 	11 , added snp: 73, loss: 11.2992, out-of-bag acc: 92.86%, # of haplo: 16
#> 	12 , added snp: 251, loss: 9.57313, out-of-bag acc: 92.86%, # of haplo: 22
#> 	13 , added snp: 212, loss: 8.52645, out-of-bag acc: 92.86%, # of haplo: 23
#> 	14 , added snp: 194, loss: 6.20003, out-of-bag acc: 92.86%, # of haplo: 26
#> 	15 , added snp: 13, loss: 5.75359, out-of-bag acc: 92.86%, # of haplo: 28
#> 	16 , added snp: 199, loss: 5.11321, out-of-bag acc: 92.86%, # of haplo: 28
#> Tue May 19 19:47:18 2026,   3 individual classifier, out-of-bag acc: 92.86%, # of SNPs: 16, # of haplo: 28
#> 	1  , added snp: 134, loss: 206.353, out-of-bag acc: 53.57%, # of haplo: 12
#> 	2  , added snp: 214, loss: 169.124, out-of-bag acc: 64.29%, # of haplo: 15
#> 	3  , added snp: 227, loss: 137.027, out-of-bag acc: 67.86%, # of haplo: 18
#> 	4  , added snp: 141, loss: 108.298, out-of-bag acc: 71.43%, # of haplo: 20
#> 	5  , added snp: 79, loss: 77.7691, out-of-bag acc: 71.43%, # of haplo: 25
#> 	6  , added snp: 140, loss: 57.6922, out-of-bag acc: 75.00%, # of haplo: 33
#> 	7  , added snp: 151, loss: 56.8582, out-of-bag acc: 78.57%, # of haplo: 34
#> 	8  , added snp: 219, loss: 44.7064, out-of-bag acc: 78.57%, # of haplo: 35
#> 	9  , added snp: 100, loss: 31.6522, out-of-bag acc: 78.57%, # of haplo: 36
#> 	10 , added snp: 160, loss: 24.2103, out-of-bag acc: 78.57%, # of haplo: 36
#> 	11 , added snp: 118, loss: 23.6526, out-of-bag acc: 78.57%, # of haplo: 36
#> 	12 , added snp: 33, loss: 23.2194, out-of-bag acc: 78.57%, # of haplo: 37
#> 	13 , added snp: 109, loss: 23.0817, out-of-bag acc: 78.57%, # of haplo: 37
#> 	14 , added snp: 199, loss: 21.2586, out-of-bag acc: 78.57%, # of haplo: 37
#> Tue May 19 19:47:18 2026,   4 individual classifier, out-of-bag acc: 78.57%, # of SNPs: 14, # of haplo: 37
summary(model)
#> Gene: HLA - A
#> Training dataset: 34 samples X 264 SNPs
#> 	# of KIR3DL1/S1 alleles: 14
#> 	# of individual classifiers: 4
#> 	total # of SNPs used: 45
#> 	average # of SNPs in an individual classifier: 14.50, sd: 1.00, min: 14, max: 16
#> 	average # of haplotypes in an individual classifier: 29.00, sd: 7.62, min: 19, max: 37
#> 	average out-of-bag accuracy: 87.39%, sd: 7.24%, min: 78.57%, max: 93.75%
#> Genome assembly: hg19

# validation
pred <- predict(model, test.geno)
#> Error in hlaPredict(object, snp, cl, type, vote, allele.check, match.type,     same.strand, verbose, verbose.match): The handle of HIBAG model has been closed.
# compare
(comp <- hlaCompareAllele(hlatab$validation, pred, allele.limit=model,
  call.threshold=0))
#> Calling 'hlaCompareAllele': there are 0 individuals in common.
#> $overall
#>   total.num.ind crt.num.ind crt.num.haplo acc.ind acc.haplo call.threshold
#> 1             0           0             0     NaN       NaN              0
#>   n.call call.rate
#> 1      0       NaN
#> 
#> $confusion
#>        True
#> Predict 01:01 02:01 02:06 03:01 11:01 23:01 24:02 24:03 25:01 26:01 29:02 31:01
#>   01:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:06     0     0     0     0     0     0     0     0     0     0     0     0
#>   03:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   11:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   23:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:03     0     0     0     0     0     0     0     0     0     0     0     0
#>   25:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   26:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   29:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   31:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   32:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   68:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   ...       0     0     0     0     0     0     0     0     0     0     0     0
#>        True
#> Predict 32:01 68:01
#>   01:01     0     0
#>   02:01     0     0
#>   02:06     0     0
#>   03:01     0     0
#>   11:01     0     0
#>   23:01     0     0
#>   24:02     0     0
#>   24:03     0     0
#>   25:01     0     0
#>   26:01     0     0
#>   29:02     0     0
#>   31:01     0     0
#>   32:01     0     0
#>   68:01     0     0
#>   ...       0     0
#> 
#> $detail
#>    allele train.num train.freq valid.num valid.freq call.rate accuracy
#> 1   01:01        13 0.19117647         0        NaN         0      NaN
#> 2   02:01        21 0.30882353         0        NaN         0      NaN
#> 3   02:06         1 0.01470588         0        NaN         0      NaN
#> 4   03:01         4 0.05882353         0        NaN         0      NaN
#> 5   11:01         3 0.04411765         0        NaN         0      NaN
#> 6   23:01         1 0.01470588         0        NaN         0      NaN
#> 7   24:02         8 0.11764706         0        NaN         0      NaN
#> 8   24:03         1 0.01470588         0        NaN         0      NaN
#> 9   25:01         4 0.05882353         0        NaN         0      NaN
#> 10  26:01         2 0.02941176         0        NaN         0      NaN
#> 11  29:02         3 0.04411765         0        NaN         0      NaN
#> 12  31:01         2 0.02941176         0        NaN         0      NaN
#> 13  32:01         3 0.04411765         0        NaN         0      NaN
#> 14  68:01         2 0.02941176         0        NaN         0      NaN
#>    sensitivity specificity ppv npv miscall miscall.prop
#> 1          NaN         NaN NaN NaN    <NA>          NaN
#> 2          NaN         NaN NaN NaN    <NA>          NaN
#> 3          NaN         NaN NaN NaN    <NA>          NaN
#> 4          NaN         NaN NaN NaN    <NA>          NaN
#> 5          NaN         NaN NaN NaN    <NA>          NaN
#> 6          NaN         NaN NaN NaN    <NA>          NaN
#> 7          NaN         NaN NaN NaN    <NA>          NaN
#> 8          NaN         NaN NaN NaN    <NA>          NaN
#> 9          NaN         NaN NaN NaN    <NA>          NaN
#> 10         NaN         NaN NaN NaN    <NA>          NaN
#> 11         NaN         NaN NaN NaN    <NA>          NaN
#> 12         NaN         NaN NaN NaN    <NA>          NaN
#> 13         NaN         NaN NaN NaN    <NA>          NaN
#> 14         NaN         NaN NaN NaN    <NA>          NaN
#> 
(comp <- hlaCompareAllele(hlatab$validation, pred, allele.limit=model,
  call.threshold=0.5))
#> Calling 'hlaCompareAllele': there are 0 individuals in common.
#> $overall
#>   total.num.ind crt.num.ind crt.num.haplo acc.ind acc.haplo call.threshold
#> 1             0           0             0     NaN       NaN            0.5
#>   n.call call.rate
#> 1      0       NaN
#> 
#> $confusion
#>        True
#> Predict 01:01 02:01 02:06 03:01 11:01 23:01 24:02 24:03 25:01 26:01 29:02 31:01
#>   01:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:06     0     0     0     0     0     0     0     0     0     0     0     0
#>   03:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   11:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   23:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:03     0     0     0     0     0     0     0     0     0     0     0     0
#>   25:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   26:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   29:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   31:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   32:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   68:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   ...       0     0     0     0     0     0     0     0     0     0     0     0
#>        True
#> Predict 32:01 68:01
#>   01:01     0     0
#>   02:01     0     0
#>   02:06     0     0
#>   03:01     0     0
#>   11:01     0     0
#>   23:01     0     0
#>   24:02     0     0
#>   24:03     0     0
#>   25:01     0     0
#>   26:01     0     0
#>   29:02     0     0
#>   31:01     0     0
#>   32:01     0     0
#>   68:01     0     0
#>   ...       0     0
#> 
#> $detail
#>    allele train.num train.freq valid.num valid.freq call.rate accuracy
#> 1   01:01        13 0.19117647         0        NaN         0      NaN
#> 2   02:01        21 0.30882353         0        NaN         0      NaN
#> 3   02:06         1 0.01470588         0        NaN         0      NaN
#> 4   03:01         4 0.05882353         0        NaN         0      NaN
#> 5   11:01         3 0.04411765         0        NaN         0      NaN
#> 6   23:01         1 0.01470588         0        NaN         0      NaN
#> 7   24:02         8 0.11764706         0        NaN         0      NaN
#> 8   24:03         1 0.01470588         0        NaN         0      NaN
#> 9   25:01         4 0.05882353         0        NaN         0      NaN
#> 10  26:01         2 0.02941176         0        NaN         0      NaN
#> 11  29:02         3 0.04411765         0        NaN         0      NaN
#> 12  31:01         2 0.02941176         0        NaN         0      NaN
#> 13  32:01         3 0.04411765         0        NaN         0      NaN
#> 14  68:01         2 0.02941176         0        NaN         0      NaN
#>    sensitivity specificity ppv npv miscall miscall.prop
#> 1          NaN         NaN NaN NaN    <NA>          NaN
#> 2          NaN         NaN NaN NaN    <NA>          NaN
#> 3          NaN         NaN NaN NaN    <NA>          NaN
#> 4          NaN         NaN NaN NaN    <NA>          NaN
#> 5          NaN         NaN NaN NaN    <NA>          NaN
#> 6          NaN         NaN NaN NaN    <NA>          NaN
#> 7          NaN         NaN NaN NaN    <NA>          NaN
#> 8          NaN         NaN NaN NaN    <NA>          NaN
#> 9          NaN         NaN NaN NaN    <NA>          NaN
#> 10         NaN         NaN NaN NaN    <NA>          NaN
#> 11         NaN         NaN NaN NaN    <NA>          NaN
#> 12         NaN         NaN NaN NaN    <NA>          NaN
#> 13         NaN         NaN NaN NaN    <NA>          NaN
#> 14         NaN         NaN NaN NaN    <NA>          NaN
#> 

# save the parameter file
mobj <- hlaModelToObj(model)
save(mobj, file="HIBAG_model.RData")
save(test.geno, file="testgeno.RData")
save(hlatab, file="HLASplit.RData")

# Clear Workspace
hlaClose(model)  # release all resources of model
rm(list = ls())

######################################################################

# NOW, load a HIBAG model from the parameter file
mobj <- get(load("HIBAG_model.RData"))
model <- hlaModelFromObj(mobj)

# validation
test.geno <- get(load("testgeno.RData"))
hlatab <- get(load("HLASplit.RData"))

pred <- predict(model, test.geno, type="response")
#> Error in hlaPredict(object, snp, cl, type, vote, allele.check, match.type,     same.strand, verbose, verbose.match): The handle of HIBAG model has been closed.
summary(pred)
#> Gene: KIR2DS5
#> Range: [55281035bp, 55296300bp] on hg19
#> # of samples: 225
#> # of unique KIR3DL1/S1 alleles: 3
#> # of unique KIR3DL1/S1 genotypes: 4
#> Posterior probability:
#>    [0,0.25)  [0.25,0.5)  [0.5,0.75)    [0.75,1] 
#>    0 (0.0%)   11 (4.9%)  97 (43.1%) 117 (52.0%) 

# compare
(comp <- hlaCompareAllele(hlatab$validation, pred, allele.limit=model,
  call.threshold=0))
#> Calling 'hlaCompareAllele': there are 0 individuals in common.
#> $overall
#>   total.num.ind crt.num.ind crt.num.haplo acc.ind acc.haplo call.threshold
#> 1             0           0             0     NaN       NaN              0
#>   n.call call.rate
#> 1      0       NaN
#> 
#> $confusion
#>        True
#> Predict 01:01 02:01 02:06 03:01 11:01 23:01 24:02 24:03 25:01 26:01 29:02 31:01
#>   01:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:06     0     0     0     0     0     0     0     0     0     0     0     0
#>   03:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   11:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   23:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:03     0     0     0     0     0     0     0     0     0     0     0     0
#>   25:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   26:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   29:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   31:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   32:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   68:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   ...       0     0     0     0     0     0     0     0     0     0     0     0
#>        True
#> Predict 32:01 68:01
#>   01:01     0     0
#>   02:01     0     0
#>   02:06     0     0
#>   03:01     0     0
#>   11:01     0     0
#>   23:01     0     0
#>   24:02     0     0
#>   24:03     0     0
#>   25:01     0     0
#>   26:01     0     0
#>   29:02     0     0
#>   31:01     0     0
#>   32:01     0     0
#>   68:01     0     0
#>   ...       0     0
#> 
#> $detail
#>    allele train.num train.freq valid.num valid.freq call.rate accuracy
#> 1   01:01        13 0.19117647         0        NaN         0      NaN
#> 2   02:01        21 0.30882353         0        NaN         0      NaN
#> 3   02:06         1 0.01470588         0        NaN         0      NaN
#> 4   03:01         4 0.05882353         0        NaN         0      NaN
#> 5   11:01         3 0.04411765         0        NaN         0      NaN
#> 6   23:01         1 0.01470588         0        NaN         0      NaN
#> 7   24:02         8 0.11764706         0        NaN         0      NaN
#> 8   24:03         1 0.01470588         0        NaN         0      NaN
#> 9   25:01         4 0.05882353         0        NaN         0      NaN
#> 10  26:01         2 0.02941176         0        NaN         0      NaN
#> 11  29:02         3 0.04411765         0        NaN         0      NaN
#> 12  31:01         2 0.02941176         0        NaN         0      NaN
#> 13  32:01         3 0.04411765         0        NaN         0      NaN
#> 14  68:01         2 0.02941176         0        NaN         0      NaN
#>    sensitivity specificity ppv npv miscall miscall.prop
#> 1          NaN         NaN NaN NaN    <NA>          NaN
#> 2          NaN         NaN NaN NaN    <NA>          NaN
#> 3          NaN         NaN NaN NaN    <NA>          NaN
#> 4          NaN         NaN NaN NaN    <NA>          NaN
#> 5          NaN         NaN NaN NaN    <NA>          NaN
#> 6          NaN         NaN NaN NaN    <NA>          NaN
#> 7          NaN         NaN NaN NaN    <NA>          NaN
#> 8          NaN         NaN NaN NaN    <NA>          NaN
#> 9          NaN         NaN NaN NaN    <NA>          NaN
#> 10         NaN         NaN NaN NaN    <NA>          NaN
#> 11         NaN         NaN NaN NaN    <NA>          NaN
#> 12         NaN         NaN NaN NaN    <NA>          NaN
#> 13         NaN         NaN NaN NaN    <NA>          NaN
#> 14         NaN         NaN NaN NaN    <NA>          NaN
#> 
(comp <- hlaCompareAllele(hlatab$validation, pred, allele.limit=model,
  call.threshold=0.5))
#> Calling 'hlaCompareAllele': there are 0 individuals in common.
#> $overall
#>   total.num.ind crt.num.ind crt.num.haplo acc.ind acc.haplo call.threshold
#> 1             0           0             0     NaN       NaN            0.5
#>   n.call call.rate
#> 1      0       NaN
#> 
#> $confusion
#>        True
#> Predict 01:01 02:01 02:06 03:01 11:01 23:01 24:02 24:03 25:01 26:01 29:02 31:01
#>   01:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:06     0     0     0     0     0     0     0     0     0     0     0     0
#>   03:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   11:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   23:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:03     0     0     0     0     0     0     0     0     0     0     0     0
#>   25:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   26:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   29:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   31:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   32:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   68:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   ...       0     0     0     0     0     0     0     0     0     0     0     0
#>        True
#> Predict 32:01 68:01
#>   01:01     0     0
#>   02:01     0     0
#>   02:06     0     0
#>   03:01     0     0
#>   11:01     0     0
#>   23:01     0     0
#>   24:02     0     0
#>   24:03     0     0
#>   25:01     0     0
#>   26:01     0     0
#>   29:02     0     0
#>   31:01     0     0
#>   32:01     0     0
#>   68:01     0     0
#>   ...       0     0
#> 
#> $detail
#>    allele train.num train.freq valid.num valid.freq call.rate accuracy
#> 1   01:01        13 0.19117647         0        NaN         0      NaN
#> 2   02:01        21 0.30882353         0        NaN         0      NaN
#> 3   02:06         1 0.01470588         0        NaN         0      NaN
#> 4   03:01         4 0.05882353         0        NaN         0      NaN
#> 5   11:01         3 0.04411765         0        NaN         0      NaN
#> 6   23:01         1 0.01470588         0        NaN         0      NaN
#> 7   24:02         8 0.11764706         0        NaN         0      NaN
#> 8   24:03         1 0.01470588         0        NaN         0      NaN
#> 9   25:01         4 0.05882353         0        NaN         0      NaN
#> 10  26:01         2 0.02941176         0        NaN         0      NaN
#> 11  29:02         3 0.04411765         0        NaN         0      NaN
#> 12  31:01         2 0.02941176         0        NaN         0      NaN
#> 13  32:01         3 0.04411765         0        NaN         0      NaN
#> 14  68:01         2 0.02941176         0        NaN         0      NaN
#>    sensitivity specificity ppv npv miscall miscall.prop
#> 1          NaN         NaN NaN NaN    <NA>          NaN
#> 2          NaN         NaN NaN NaN    <NA>          NaN
#> 3          NaN         NaN NaN NaN    <NA>          NaN
#> 4          NaN         NaN NaN NaN    <NA>          NaN
#> 5          NaN         NaN NaN NaN    <NA>          NaN
#> 6          NaN         NaN NaN NaN    <NA>          NaN
#> 7          NaN         NaN NaN NaN    <NA>          NaN
#> 8          NaN         NaN NaN NaN    <NA>          NaN
#> 9          NaN         NaN NaN NaN    <NA>          NaN
#> 10         NaN         NaN NaN NaN    <NA>          NaN
#> 11         NaN         NaN NaN NaN    <NA>          NaN
#> 12         NaN         NaN NaN NaN    <NA>          NaN
#> 13         NaN         NaN NaN NaN    <NA>          NaN
#> 14         NaN         NaN NaN NaN    <NA>          NaN
#>