Merge prediction results from multiple HIBAG models

Return an object of hlaAlleleClass, which contains predicted HLA types.

hlaPredMerge(..., weight=NULL, equivalence=NULL)

Arguments

...

The object(s) of hlaAlleleClass, having a field of 'postprob', and returned by predict(..., type="response+prob", vote="majority")

weight

the weight used for each prediction; if NULL, equal weights

equivalence

a data.frame with two columns, the first column for new equivalent alleles, and the second for the alleles possibly existed in the object(s) passed to this function

Details

Calculate a new probability matrix for each pair of HLA alleles, by averaging (posterior) probabilities from all models with specified weights. If equivalence is specified, multiple alleles might be collapsed into one class.

Value

Return a hlaAlleleClass object.

Author

Xiuwen Zheng

See also

Examples

# load HLA types and SNP genotypes
data(HLA_Type_Table, package="HIBAG")
data(HapMap_CEU_Geno, package="HIBAG")

# make a "hlaAlleleClass" object
hla.id <- "A"
hla <- hlaAllele(HLA_Type_Table$sample.id,
  H1 = HLA_Type_Table[, paste(hla.id, ".1", sep="")],
  H2 = HLA_Type_Table[, paste(hla.id, ".2", sep="")],
  locus=hla.id, assembly="hg19")

# divide HLA types randomly
set.seed(100)
hlatab <- hlaSplitAllele(hla, train.prop=0.5)
names(hlatab)
#> [1] "training"   "validation"
# "training"   "validation"
summary(hlatab$training)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> # of unique KIR3DL1/S1 genotypes: 21
summary(hlatab$validation)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 26
#> # of unique KIR3DL1/S1 alleles: 12
#> # of unique KIR3DL1/S1 genotypes: 17

# SNP predictors within the flanking region on each side
region <- 500   # kb
snpid <- hlaFlankingSNP(HapMap_CEU_Geno$snp.id, HapMap_CEU_Geno$snp.position,
  hla.id, region*1000, assembly="hg19")
length(snpid)  # 275
#> [1] 275

# training and validation genotypes
train.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  snp.sel=match(snpid, HapMap_CEU_Geno$snp.id),
  samp.sel=match(hlatab$training$value$sample.id, HapMap_CEU_Geno$sample.id))
test.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  samp.sel=match(hlatab$validation$value$sample.id, HapMap_CEU_Geno$sample.id))

# train HIBAG models
set.seed(100)

# please use "nclassifier=100" when you use HIBAG for real data
m1 <- hlaAttrBagging(hlatab$training, train.geno, nclassifier=2,
  verbose.detail=TRUE)
#> 11 monomorphic SNPs have been removed.
#> Build a PONG2 model with 2 individual classifiers:
#> # of SNPs randomly sampled as candidates for each selection: 17
#> # of SNPs: 264, # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> 	1  , added snp: 102, loss: 149.075, out-of-bag acc: 53.57%, # of haplo: 11
#> 	2  , added snp: 136, loss: 120.694, out-of-bag acc: 64.29%, # of haplo: 11
#> 	3  , added snp: 132, loss: 83.1029, out-of-bag acc: 71.43%, # of haplo: 11
#> 	4  , added snp: 158, loss: 61.1726, out-of-bag acc: 75.00%, # of haplo: 11
#> 	5  , added snp: 130, loss: 22.8538, out-of-bag acc: 78.57%, # of haplo: 11
#> 	6  , added snp: 107, loss: 5.32632, out-of-bag acc: 82.14%, # of haplo: 11
#> 	7  , added snp: 75, loss: 3.91769, out-of-bag acc: 82.14%, # of haplo: 13
#> 	8  , added snp: 241, loss: 3.82069, out-of-bag acc: 82.14%, # of haplo: 14
#> 	9  , added snp: 248, loss: 0.00121148, out-of-bag acc: 82.14%, # of haplo: 16
#> Tue May 19 19:47:32 2026,   1 individual classifier, out-of-bag acc: 82.14%, # of SNPs: 9, # of haplo: 16
#> 	1  , added snp: 122, loss: 146.45, out-of-bag acc: 50.00%, # of haplo: 11
#> 	2  , added snp: 134, loss: 107.665, out-of-bag acc: 60.71%, # of haplo: 11
#> 	3  , added snp: 129, loss: 60.1399, out-of-bag acc: 67.86%, # of haplo: 11
#> 	4  , added snp: 102, loss: 32.5122, out-of-bag acc: 78.57%, # of haplo: 11
#> 	5  , added snp: 82, loss: 25.4615, out-of-bag acc: 82.14%, # of haplo: 11
#> 	6  , added snp: 150, loss: 9.73285, out-of-bag acc: 85.71%, # of haplo: 11
#> 	7  , added snp: 100, loss: 0.00161299, out-of-bag acc: 85.71%, # of haplo: 11
#> Tue May 19 19:47:32 2026,   2 individual classifier, out-of-bag acc: 85.71%, # of SNPs: 7, # of haplo: 11
m2 <- hlaAttrBagging(hlatab$training, train.geno, nclassifier=2,
  verbose.detail=TRUE)
#> 11 monomorphic SNPs have been removed.
#> Build a PONG2 model with 2 individual classifiers:
#> # of SNPs randomly sampled as candidates for each selection: 17
#> # of SNPs: 264, # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> 	1  , added snp: 158, loss: 194.579, out-of-bag acc: 54.55%, # of haplo: 16
#> 	2  , added snp: 121, loss: 143.987, out-of-bag acc: 63.64%, # of haplo: 17
#> 	3  , added snp: 129, loss: 106.506, out-of-bag acc: 77.27%, # of haplo: 18
#> 	4  , added snp: 77, loss: 60.9406, out-of-bag acc: 81.82%, # of haplo: 21
#> 	5  , added snp: 212, loss: 45.4935, out-of-bag acc: 86.36%, # of haplo: 21
#> 	6  , added snp: 63, loss: 27.9572, out-of-bag acc: 90.91%, # of haplo: 21
#> 	7  , added snp: 210, loss: 21.1363, out-of-bag acc: 90.91%, # of haplo: 22
#> 	8  , added snp: 132, loss: 16.9832, out-of-bag acc: 90.91%, # of haplo: 22
#> 	9  , added snp: 194, loss: 16.0641, out-of-bag acc: 90.91%, # of haplo: 26
#> 	10 , added snp: 16, loss: 13.9943, out-of-bag acc: 90.91%, # of haplo: 36
#> 	11 , added snp: 105, loss: 13.8524, out-of-bag acc: 90.91%, # of haplo: 36
#> 	12 , added snp: 199, loss: 12.5987, out-of-bag acc: 90.91%, # of haplo: 36
#> 	13 , added snp: 208, loss: 10.448, out-of-bag acc: 90.91%, # of haplo: 37
#> 	14 , added snp: 191, loss: 10.0943, out-of-bag acc: 90.91%, # of haplo: 37
#> Tue May 19 19:47:32 2026,   1 individual classifier, out-of-bag acc: 90.91%, # of SNPs: 14, # of haplo: 37
#> 	1  , added snp: 127, loss: 212.355, out-of-bag acc: 72.22%, # of haplo: 15
#> 	2  , added snp: 117, loss: 145.94, out-of-bag acc: 83.33%, # of haplo: 17
#> 	3  , added snp: 218, loss: 112.652, out-of-bag acc: 88.89%, # of haplo: 20
#> 	4  , added snp: 123, loss: 87.4831, out-of-bag acc: 94.44%, # of haplo: 22
#> 	5  , added snp: 184, loss: 76.6015, out-of-bag acc: 100.00%, # of haplo: 27
#> 	6  , added snp: 216, loss: 59.7243, out-of-bag acc: 100.00%, # of haplo: 30
#> 	7  , added snp: 205, loss: 46.2989, out-of-bag acc: 100.00%, # of haplo: 35
#> 	8  , added snp: 233, loss: 36.6235, out-of-bag acc: 100.00%, # of haplo: 36
#> 	9  , added snp: 163, loss: 31.6074, out-of-bag acc: 100.00%, # of haplo: 37
#> 	10 , added snp: 94, loss: 29.6891, out-of-bag acc: 100.00%, # of haplo: 38
#> 	11 , added snp: 3, loss: 25.5841, out-of-bag acc: 100.00%, # of haplo: 41
#> 	12 , added snp: 58, loss: 24.2, out-of-bag acc: 100.00%, # of haplo: 45
#> 	13 , added snp: 208, loss: 23.2387, out-of-bag acc: 100.00%, # of haplo: 49
#> 	14 , added snp: 8, loss: 20.5705, out-of-bag acc: 100.00%, # of haplo: 49
#> 	15 , added snp: 151, loss: 19.0685, out-of-bag acc: 100.00%, # of haplo: 49
#> 	16 , added snp: 199, loss: 17.8042, out-of-bag acc: 100.00%, # of haplo: 53
#> 	17 , added snp: 73, loss: 13.4541, out-of-bag acc: 100.00%, # of haplo: 62
#> Tue May 19 19:47:32 2026,   2 individual classifier, out-of-bag acc: 100.00%, # of SNPs: 17, # of haplo: 62

# validation
pd1 <- predict(m1, test.geno, type="response+prob", vote="majority")
#> Error in hlaPredict(object, snp, cl, type, vote, allele.check, match.type,     same.strand, verbose, verbose.match): The handle of HIBAG model has been closed.
pd2 <- predict(m2, test.geno, type="response+prob", vote="majority")
#> Error in hlaPredict(object, snp, cl, type, vote, allele.check, match.type,     same.strand, verbose, verbose.match): The handle of HIBAG model has been closed.

hlaCompareAllele(hlatab$validation, pd1)$overall
#> Error: object 'pd1' not found
# acc.haplo = 0.8269231

hlaCompareAllele(hlatab$validation, pd2)$overall
#> Error: object 'pd2' not found
# acc.haplo = 0.8076923

# merge predictions from multiple models, by voting from all classifiers
pd <- hlaPredMerge(pd1, pd2, weight=c(1,1))
#> Error: object 'pd1' not found

hlaCompareAllele(hlatab$validation, pd)$overall
#> Error: object 'pd' not found
# acc.haplo = 0.9230769