Format a report

Create a report for evaluating prediction accuracies.

hlaReport(object, export.fn="", type=c("txt", "tex", "html"),
  header=TRUE)

Arguments

object

an object returned by hlaCompareAllele

export.fn

a file name for output, or "" for stdout

type

"txt" – tab-delimited text format; "tex" – tex format using the 'longtable' package; "html" – html file

header

if TRUE, output the header of text file associated corresponding format

Value

None.

Author

Xiuwen Zheng

See also

Examples

# load HLA types and SNP genotypes
data(HLA_Type_Table, package="HIBAG")
data(HapMap_CEU_Geno, package="HIBAG")

# make a "hlaAlleleClass" object
hla.id <- "A"
hla <- hlaAllele(HLA_Type_Table$sample.id,
  H1 = HLA_Type_Table[, paste(hla.id, ".1", sep="")],
  H2 = HLA_Type_Table[, paste(hla.id, ".2", sep="")],
  locus=hla.id, assembly="hg19")

# divide HLA types randomly
set.seed(100)
hlatab <- hlaSplitAllele(hla, train.prop=0.5)
names(hlatab)
#> [1] "training"   "validation"
# "training"   "validation"
summary(hlatab$training)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> # of unique KIR3DL1/S1 genotypes: 21
summary(hlatab$validation)
#> Gene: HLA - A
#> Range: [NAbp, NAbp]
#> # of samples: 26
#> # of unique KIR3DL1/S1 alleles: 12
#> # of unique KIR3DL1/S1 genotypes: 17

# SNP predictors within the flanking region on each side
region <- 500   # kb
snpid <- hlaFlankingSNP(HapMap_CEU_Geno$snp.id, HapMap_CEU_Geno$snp.position,
  hla.id, region*1000, assembly="hg19")
length(snpid)  # 275
#> [1] 275

# training and validation genotypes
train.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  snp.sel = match(snpid, HapMap_CEU_Geno$snp.id),
  samp.sel = match(hlatab$training$value$sample.id, HapMap_CEU_Geno$sample.id))
test.geno <- hlaGenoSubset(HapMap_CEU_Geno,
  samp.sel=match(hlatab$validation$value$sample.id, HapMap_CEU_Geno$sample.id))

# train a HIBAG model
set.seed(100)
# please use "nclassifier=100" when you use HIBAG for real data
model <- hlaAttrBagging(hlatab$training, train.geno, nclassifier=4,
  verbose.detail=TRUE)
#> 11 monomorphic SNPs have been removed.
#> Build a PONG2 model with 4 individual classifiers:
#> # of SNPs randomly sampled as candidates for each selection: 17
#> # of SNPs: 264, # of samples: 34
#> # of unique KIR3DL1/S1 alleles: 14
#> 	1  , added snp: 102, loss: 149.075, out-of-bag acc: 53.57%, # of haplo: 11
#> 	2  , added snp: 136, loss: 120.694, out-of-bag acc: 64.29%, # of haplo: 11
#> 	3  , added snp: 132, loss: 83.1029, out-of-bag acc: 71.43%, # of haplo: 11
#> 	4  , added snp: 158, loss: 61.1726, out-of-bag acc: 75.00%, # of haplo: 11
#> 	5  , added snp: 130, loss: 22.8538, out-of-bag acc: 78.57%, # of haplo: 11
#> 	6  , added snp: 107, loss: 5.32632, out-of-bag acc: 82.14%, # of haplo: 11
#> 	7  , added snp: 75, loss: 3.91769, out-of-bag acc: 82.14%, # of haplo: 13
#> 	8  , added snp: 241, loss: 3.82069, out-of-bag acc: 82.14%, # of haplo: 14
#> 	9  , added snp: 248, loss: 0.00121148, out-of-bag acc: 82.14%, # of haplo: 16
#> Tue May 19 19:47:33 2026,   1 individual classifier, out-of-bag acc: 82.14%, # of SNPs: 9, # of haplo: 16
#> 	1  , added snp: 122, loss: 146.45, out-of-bag acc: 50.00%, # of haplo: 11
#> 	2  , added snp: 134, loss: 107.665, out-of-bag acc: 60.71%, # of haplo: 11
#> 	3  , added snp: 129, loss: 60.1399, out-of-bag acc: 67.86%, # of haplo: 11
#> 	4  , added snp: 102, loss: 32.5122, out-of-bag acc: 78.57%, # of haplo: 11
#> 	5  , added snp: 82, loss: 25.4615, out-of-bag acc: 82.14%, # of haplo: 11
#> 	6  , added snp: 150, loss: 9.73285, out-of-bag acc: 85.71%, # of haplo: 11
#> 	7  , added snp: 100, loss: 0.00161299, out-of-bag acc: 85.71%, # of haplo: 11
#> Tue May 19 19:47:33 2026,   2 individual classifier, out-of-bag acc: 85.71%, # of SNPs: 7, # of haplo: 11
#> 	1  , added snp: 158, loss: 194.579, out-of-bag acc: 54.55%, # of haplo: 16
#> 	2  , added snp: 121, loss: 143.987, out-of-bag acc: 63.64%, # of haplo: 17
#> 	3  , added snp: 129, loss: 106.506, out-of-bag acc: 77.27%, # of haplo: 18
#> 	4  , added snp: 77, loss: 60.9406, out-of-bag acc: 81.82%, # of haplo: 21
#> 	5  , added snp: 212, loss: 45.4935, out-of-bag acc: 86.36%, # of haplo: 21
#> 	6  , added snp: 63, loss: 27.9572, out-of-bag acc: 90.91%, # of haplo: 21
#> 	7  , added snp: 210, loss: 21.1363, out-of-bag acc: 90.91%, # of haplo: 22
#> 	8  , added snp: 132, loss: 16.9832, out-of-bag acc: 90.91%, # of haplo: 22
#> 	9  , added snp: 194, loss: 16.0641, out-of-bag acc: 90.91%, # of haplo: 26
#> 	10 , added snp: 16, loss: 13.9943, out-of-bag acc: 90.91%, # of haplo: 36
#> 	11 , added snp: 105, loss: 13.8524, out-of-bag acc: 90.91%, # of haplo: 36
#> 	12 , added snp: 199, loss: 12.5987, out-of-bag acc: 90.91%, # of haplo: 36
#> 	13 , added snp: 208, loss: 10.448, out-of-bag acc: 90.91%, # of haplo: 37
#> 	14 , added snp: 191, loss: 10.0943, out-of-bag acc: 90.91%, # of haplo: 37
#> Tue May 19 19:47:33 2026,   3 individual classifier, out-of-bag acc: 90.91%, # of SNPs: 14, # of haplo: 37
#> 	1  , added snp: 127, loss: 212.355, out-of-bag acc: 72.22%, # of haplo: 15
#> 	2  , added snp: 117, loss: 145.94, out-of-bag acc: 83.33%, # of haplo: 17
#> 	3  , added snp: 218, loss: 112.652, out-of-bag acc: 88.89%, # of haplo: 20
#> 	4  , added snp: 123, loss: 87.4831, out-of-bag acc: 94.44%, # of haplo: 22
#> 	5  , added snp: 184, loss: 76.6015, out-of-bag acc: 100.00%, # of haplo: 27
#> 	6  , added snp: 216, loss: 59.7243, out-of-bag acc: 100.00%, # of haplo: 30
#> 	7  , added snp: 205, loss: 46.2989, out-of-bag acc: 100.00%, # of haplo: 35
#> 	8  , added snp: 233, loss: 36.6235, out-of-bag acc: 100.00%, # of haplo: 36
#> 	9  , added snp: 163, loss: 31.6074, out-of-bag acc: 100.00%, # of haplo: 37
#> 	10 , added snp: 94, loss: 29.6891, out-of-bag acc: 100.00%, # of haplo: 38
#> 	11 , added snp: 3, loss: 25.5841, out-of-bag acc: 100.00%, # of haplo: 41
#> 	12 , added snp: 58, loss: 24.2, out-of-bag acc: 100.00%, # of haplo: 45
#> 	13 , added snp: 208, loss: 23.2387, out-of-bag acc: 100.00%, # of haplo: 49
#> 	14 , added snp: 8, loss: 20.5705, out-of-bag acc: 100.00%, # of haplo: 49
#> 	15 , added snp: 151, loss: 19.0685, out-of-bag acc: 100.00%, # of haplo: 49
#> 	16 , added snp: 199, loss: 17.8042, out-of-bag acc: 100.00%, # of haplo: 53
#> 	17 , added snp: 73, loss: 13.4541, out-of-bag acc: 100.00%, # of haplo: 62
#> Tue May 19 19:47:33 2026,   4 individual classifier, out-of-bag acc: 100.00%, # of SNPs: 17, # of haplo: 62
summary(model)
#> Gene: HLA - A
#> Training dataset: 34 samples X 264 SNPs
#> 	# of KIR3DL1/S1 alleles: 14
#> 	# of individual classifiers: 4
#> 	total # of SNPs used: 41
#> 	average # of SNPs in an individual classifier: 11.75, sd: 4.57, min: 7, max: 17
#> 	average # of haplotypes in an individual classifier: 31.50, sd: 23.25, min: 11, max: 62
#> 	average out-of-bag accuracy: 89.69%, sd: 7.76%, min: 82.14%, max: 100.00%
#> Genome assembly: hg19

# validation
pred <- predict(model, test.geno)
#> Error in hlaPredict(object, snp, cl, type, vote, allele.check, match.type,     same.strand, verbose, verbose.match): The handle of HIBAG model has been closed.
# compare
(comp <- hlaCompareAllele(hlatab$validation, pred, allele.limit=model,
  call.threshold=0))
#> Calling 'hlaCompareAllele': there are 0 individuals in common.
#> $overall
#>   total.num.ind crt.num.ind crt.num.haplo acc.ind acc.haplo call.threshold
#> 1             0           0             0     NaN       NaN              0
#>   n.call call.rate
#> 1      0       NaN
#> 
#> $confusion
#>        True
#> Predict 01:01 02:01 02:06 03:01 11:01 23:01 24:02 24:03 25:01 26:01 29:02 31:01
#>   01:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   02:06     0     0     0     0     0     0     0     0     0     0     0     0
#>   03:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   11:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   23:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   24:03     0     0     0     0     0     0     0     0     0     0     0     0
#>   25:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   26:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   29:02     0     0     0     0     0     0     0     0     0     0     0     0
#>   31:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   32:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   68:01     0     0     0     0     0     0     0     0     0     0     0     0
#>   ...       0     0     0     0     0     0     0     0     0     0     0     0
#>        True
#> Predict 32:01 68:01
#>   01:01     0     0
#>   02:01     0     0
#>   02:06     0     0
#>   03:01     0     0
#>   11:01     0     0
#>   23:01     0     0
#>   24:02     0     0
#>   24:03     0     0
#>   25:01     0     0
#>   26:01     0     0
#>   29:02     0     0
#>   31:01     0     0
#>   32:01     0     0
#>   68:01     0     0
#>   ...       0     0
#> 
#> $detail
#>    allele train.num train.freq valid.num valid.freq call.rate accuracy
#> 1   01:01        13 0.19117647         0        NaN         0      NaN
#> 2   02:01        25 0.36764706         0        NaN         0      NaN
#> 3   02:06         1 0.01470588         0        NaN         0      NaN
#> 4   03:01         4 0.05882353         0        NaN         0      NaN
#> 5   11:01         2 0.02941176         0        NaN         0      NaN
#> 6   23:01         1 0.01470588         0        NaN         0      NaN
#> 7   24:02         6 0.08823529         0        NaN         0      NaN
#> 8   24:03         1 0.01470588         0        NaN         0      NaN
#> 9   25:01         4 0.05882353         0        NaN         0      NaN
#> 10  26:01         2 0.02941176         0        NaN         0      NaN
#> 11  29:02         3 0.04411765         0        NaN         0      NaN
#> 12  31:01         2 0.02941176         0        NaN         0      NaN
#> 13  32:01         2 0.02941176         0        NaN         0      NaN
#> 14  68:01         2 0.02941176         0        NaN         0      NaN
#>    sensitivity specificity ppv npv miscall miscall.prop
#> 1          NaN         NaN NaN NaN    <NA>          NaN
#> 2          NaN         NaN NaN NaN    <NA>          NaN
#> 3          NaN         NaN NaN NaN    <NA>          NaN
#> 4          NaN         NaN NaN NaN    <NA>          NaN
#> 5          NaN         NaN NaN NaN    <NA>          NaN
#> 6          NaN         NaN NaN NaN    <NA>          NaN
#> 7          NaN         NaN NaN NaN    <NA>          NaN
#> 8          NaN         NaN NaN NaN    <NA>          NaN
#> 9          NaN         NaN NaN NaN    <NA>          NaN
#> 10         NaN         NaN NaN NaN    <NA>          NaN
#> 11         NaN         NaN NaN NaN    <NA>          NaN
#> 12         NaN         NaN NaN NaN    <NA>          NaN
#> 13         NaN         NaN NaN NaN    <NA>          NaN
#> 14         NaN         NaN NaN NaN    <NA>          NaN
#> 

# report
hlaReport(comp, type="txt")
#> Allele	Num.	Freq.	Num.	Freq.	CR	ACC	SEN	SPE	PPV	NPV	Miscall
#> 	Train	Train	Valid.	Valid.	(%)	(%)	(%)	(%)	(%)	(%)	(%)
#> ----
#> Overall accuracy: NaN%, Call rate: NaN%
#> 01:01	13	0.1912	0	0	--	--	--	--	--	--	--
#> 02:01	25	0.3676	0	0	--	--	--	--	--	--	--
#> 02:06	1	0.0147	0	0	--	--	--	--	--	--	--
#> 03:01	4	0.0588	0	0	--	--	--	--	--	--	--
#> 11:01	2	0.0294	0	0	--	--	--	--	--	--	--
#> 23:01	1	0.0147	0	0	--	--	--	--	--	--	--
#> 24:02	6	0.0882	0	0	--	--	--	--	--	--	--
#> 24:03	1	0.0147	0	0	--	--	--	--	--	--	--
#> 25:01	4	0.0588	0	0	--	--	--	--	--	--	--
#> 26:01	2	0.0294	0	0	--	--	--	--	--	--	--
#> 29:02	3	0.0441	0	0	--	--	--	--	--	--	--
#> 31:01	2	0.0294	0	0	--	--	--	--	--	--	--
#> 32:01	2	0.0294	0	0	--	--	--	--	--	--	--
#> 68:01	2	0.0294	0	0	--	--	--	--	--	--	--

hlaReport(comp, type="tex")
#> \title{Imputation Evaluation}
#> 
#> \documentclass[12pt]{article}
#> 
#> \usepackage{fullpage}
#> \usepackage{longtable}
#> 
#> \begin{document}
#> 
#> \maketitle
#> 
#> \setlength{\LTcapwidth}{6.4in}
#> 
#> % -------- BEGIN TABLE --------
#> \begin{longtable}{rrrrr | rrrrrrl}
#> \caption{The sensitivity (SEN), specificity (SPE), positive predictive value (PPV), negative predictive value (NPV) and call rate (CR)}
#> \label{tab:accuracy} \\
#> Allele & Num. & Freq. & Num. & Freq. & CR & ACC & SEN & SPE & PPV & NPV & Miscall \\
#>  & Train & Train & Valid. & Valid. & (\%) & (\%) & (\%) & (\%) & (\%) & (\%) & (\%) \\
#> \hline\hline
#> \endfirsthead
#> \multicolumn{12}{c}{{\normalsize \tablename\ \thetable{} -- Continued from previous page}} \\
#> Allele & Num. & Freq. & Num. & Freq. & CR & ACC & SEN & SPE & PPV & NPV & Miscall \\
#>  & Train & Train & Valid. & Valid. & (\%) & (\%) & (\%) & (\%) & (\%) & (\%) & (\%) \\
#> \hline\hline
#> \endhead
#> \hline
#> \multicolumn{12}{r}{Continued on next page ...} \\
#> \hline
#> \endfoot
#> \hline\hline
#> \endlastfoot
#> \multicolumn{12}{l}{\it Overall accuracy: NaN\%, Call rate: NaN\%} \\
#> 01:01 & 13 & 0.1912 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 02:01 & 25 & 0.3676 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 02:06 & 1 & 0.0147 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 03:01 & 4 & 0.0588 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 11:01 & 2 & 0.0294 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 23:01 & 1 & 0.0147 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 24:02 & 6 & 0.0882 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 24:03 & 1 & 0.0147 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 25:01 & 4 & 0.0588 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 26:01 & 2 & 0.0294 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 29:02 & 3 & 0.0441 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 31:01 & 2 & 0.0294 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 32:01 & 2 & 0.0294 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> 68:01 & 2 & 0.0294 & 0 & 0 & -- & -- & -- & -- & -- & -- & -- \\
#> \end{longtable}
#> % -------- END TABLE --------
#> 
#> \end{document}

hlaReport(comp, type="html")
#> <!DOCTYPE html>
#> <html>
#> <head>
#>   <title>Imputation Evaluation</title>
#> </head>
#> <body>
#> <h1>Imputation Evaluation</h1>
#> <p></p>
#> <h3><b>Table 1:</b> The sensitivity (SEN), specificity (SPE), positive predictive value (PPV), negative predictive value (NPV) and call rate (CR).</h3>
#> <table id="TB-Acc" class="tabular" border="1"  CELLSPACING="1">
#> <tr>
#> <th>Allele </th> <th>Num. Train</th> <th>Freq. Train</th> <th>Num. Valid.</th> <th>Freq. Valid.</th> <th>CR (%)</th> <th>ACC (%)</th> <th>SEN (%)</th> <th>SPE (%)</th> <th>PPV (%)</th> <th>NPV (%)</th> <th>Miscall (%)</th>
#> </tr>
#> <tr>
#> <td colspan="12">
#> <i> Overall accuracy: NaN%, Call rate: NaN% </i>
#> </td>
#> </tr>
#> <tr>
#> <td>01:01</td> <td>13</td> <td>0.1912</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>02:01</td> <td>25</td> <td>0.3676</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>02:06</td> <td>1</td> <td>0.0147</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>03:01</td> <td>4</td> <td>0.0588</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>11:01</td> <td>2</td> <td>0.0294</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>23:01</td> <td>1</td> <td>0.0147</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>24:02</td> <td>6</td> <td>0.0882</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>24:03</td> <td>1</td> <td>0.0147</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>25:01</td> <td>4</td> <td>0.0588</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>26:01</td> <td>2</td> <td>0.0294</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>29:02</td> <td>3</td> <td>0.0441</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>31:01</td> <td>2</td> <td>0.0294</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>32:01</td> <td>2</td> <td>0.0294</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> <tr>
#> <td>68:01</td> <td>2</td> <td>0.0294</td> <td>0</td> <td>0</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td> <td>--</td>
#> </tr>
#> </table>
#> 
#> </body>
#> </html>